requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
Flask==1.1.1
google-auth==1.11.2
numpy==1.18.1
pandas==1.0.1
python-telegram-bot==12.4.2
requests==2.7.0
tensorflow==2.1.0